change the landingpage with index.js
