+++ components/CheckButton/CheckButton.js
@@ -14,6 +14,7 @@ const StyledCheckButton = styled.button`
   align-self: center;
   color: white;
   transition: 0.6s;
+  cursor: pointer;
   &:hover {
     background-color: white;
   }
