+++ components/CheckButton/GreenHostsButton.js
@@ -15,6 +15,7 @@ const StyledGreenHostsButton = styled.button`
   align-self: center;
   color: white;
   transition: 0.6s;
+  cursor: pointer;
   &:hover {
     background-color: white;
   }
